pyproject.toml
[project]
name = "uv-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "datauri>=1.0.0",
    "mistralai>=1.7.0",
    "python-dotenv>=1.1.0",
]
